cMessage changes: length changed to int64 (was long); kind and priority changed to short (was int)

+++ src/sim/cbasicchannel.cc
@@ -144,7 +144,7 @@ bool cBasicChannel::deliver(cMessage *msg, simtime_t t)
     // bit error rate modeling
     if (flags & FL_ERROR_NONZERO)
     {
-        if (dblrand() < 1.0 - pow(1.0-error_, msg->length()))
+        if (dblrand() < 1.0 - pow(1.0-error_, (double)msg->length()))
             msg->setBitError(true);
     }
